[Chores #159977976] Setup the data and all necessary connections [Finishes #159977976]
