+++ package.json
@@ -1,47 +1,47 @@
 {
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "Andela Challenge 2",
-  "engines": {
-    "node": "v8.11.1",
-    "npm": "6.4.0"
-  },
-  "main": "index.js",
-  "scripts": {
-    "start": "babel-node ./server/index.js --exec babel-node --preset es2015",
-    "migrate": "node -r babel-register -r ./node_modules/dotenv/config ./server/models",
-    "babel": "babel --preset es2015 ./test/index.spec.js -o ./dist/test.spec.js",
-    "clean": "rm -rf dist",
-    "build-dev": "node bin/dev",
-    "build-dist": "npm run clean && mkdir dist && babel ./server -s -d dist",
-    "test": "mocha --require babel-core/register ./test --recursive --exit",
-    "coverage": "istanbul cover ./node_modules/mocha/bin/_mocha test/*.js",
-    "cover": "istanbul cover _mocha",
-    "report": "nyc --reporter=html --reporter=text-lcov mocha --require babel-core/register ./test --exit | coveralls",
-    "coveralls": "npm run cover -- --report lcovonly && cat ./coverage/lcov.info | coveralls"
-  },
-  "author": "Ajima Chukwuemeka",
-  "license": "ISC",
-  "dependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-istanbul": "^0.12.2",
-    "babel-preset-es2015": "^6.24.1",
-    "body-parser": "^1.18.3",
-    "chai-http": "^4.0.0",
-    "express": "^4.16.3",
-    "nyc": "^12.0.2"
-  },
-  "devDependencies": {
-    "babel-preset-stage-2": "^6.24.1",
-    "babel-register": "^6.26.0",
-    "chai": "^4.1.2",
-    "coveralls": "^3.0.2",
-    "eslint": "^5.3.0",
-    "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-plugin-import": "^2.14.0",
-    "istanbul": "^0.4.5",
-    "mocha": "^5.2.0",
-    "mocha-lcov-reporter": "^1.3.0"
-  }
+    "name": "backend",
+    "version": "1.0.0",
+    "description": "Andela Challenge 2",
+    "engines": {
+        "node": "v8.11.1",
+        "npm": "6.4.0"
+    },
+    "main": "index.js",
+    "scripts": {
+        "start": "babel-node ./server/index.js --exec babel-node --preset es2015",
+        "migrate": "node -r babel-register -r ./node_modules/dotenv/config ./server/models",
+        "babel": "babel --preset es2015 ./test/index.spec.js -o ./dist/test.spec.js",
+        "clean": "rm -rf dist",
+        "build-dev": "node bin/dev",
+        "build-dist": "npm run clean && mkdir dist && babel ./server -s -d dist",
+        "test": "mocha --require babel-core/register ./test --recursive --exit",
+        "coverage": "istanbul cover ./node_modules/mocha/bin/_mocha test/*.js",
+        "cover": "istanbul cover _mocha",
+        "report": "nyc --reporter=html --reporter=text-lcov mocha --require babel-core/register ./test --exit | coveralls",
+        "coveralls": "npm run cover -- --report lcovonly && cat ./coverage/lcov.info | coveralls"
+    },
+    "author": "Ajima Chukwuemeka",
+    "license": "ISC",
+    "dependencies": {
+        "babel-cli": "^6.26.0",
+        "babel-core": "^6.26.3",
+        "babel-istanbul": "^0.12.2",
+        "babel-preset-es2015": "^6.24.1",
+        "body-parser": "^1.18.3",
+        "chai-http": "^4.0.0",
+        "express": "^4.16.3",
+        "nyc": "^12.0.2"
+    },
+    "devDependencies": {
+        "babel-preset-stage-2": "^6.24.1",
+        "babel-register": "^6.26.0",
+        "chai": "^4.1.2",
+        "coveralls": "^3.0.2",
+        "eslint": "^5.3.0",
+        "eslint-config-airbnb-base": "^13.1.0",
+        "eslint-plugin-import": "^2.14.0",
+        "istanbul": "^0.4.5",
+        "mocha": "^5.2.0",
+        "mocha-lcov-reporter": "^1.3.0"
+    }
 }
