+++ server/controllers/datastructures/HomeController.js
@@ -1,0 +1,9 @@
+class UserController {
+    home(req, res) {
+        res.status(200).json({
+            Message: 'Welcome to StackoverFlowLite',
+        });
+    }
+}
+
+export default new UserController();
